varco-vision/requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
av==15.0.0
bitsandbytes==0.41.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
datasets==2.16.1
decord==0.6.0
deepspeed==0.14.4
dill==0.3.7
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
fastapi==0.115.14
filelock==3.18.0
flash-attn==2.8.0.post2
frozenlist==1.7.0
fsspec==2023.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gradio_client==0.2.9
h11==0.14.0
hf-xet==1.1.5
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
latex2mathml==3.78.0
-e git+https://github.com/LLaVA-VL/LLaVA-NeXT@b42941ceba259d5df18f8df8193a3897296a0449#egg=llava
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.15
networkx==3.5
ninja==1.11.1.4
numpy==1.26.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.32.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
peft==0.4.0
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pydantic==1.10.8
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.16.0
sentencepiece==0.1.99
sentry-sdk==2.32.0
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
svgwrite==1.4.3
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.16
tokenizers==0.15.2
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers.git@1c39974a4c4036fd641bc1191cc32799f85715a4
triton==3.3.1
typeguard==4.4.4
typing_extensions==4.14.0
tyro==0.9.26
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.35.0
wandb==0.21.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
